requirements.txt
flask==2.3.3
pyyaml==6.0.1
requests==2.31.0
watchdog==3.0.0
python-dotenv==1.0.0
